% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMP.R
\name{JMP}
\alias{JMP}
\title{Joint Model for Palliative care studies}
\usage{
JMP(dat = NA, dataPath = NA, covariate.fields, qol.prefix,
  qol.time.prefix, id.field, survival.time.field, censoring.status.field,
  treatment.status.field, long.knot.nums = 2:11)
}
\arguments{
\item{dat}{formatted R dataset. See example dataset: data(dat_JMP)}

\item{dataPath}{path of the data set to read in (if dat not supplied)}

\item{covariate.fields}{covariates adjusted for}

\item{qol.prefix}{prefix of the longitudinal quality of life measurements}

\item{qol.time.prefix}{prefix of the time variable for longitudinal quality of life measurements}

\item{id.field}{id variable in the dataset}

\item{survival.time.field}{survival time variable in the dataset}

\item{censoring.status.field}{censoring variable (1 for death, 0 for censored)}

\item{treatment.status.field}{treatment status (1 for treated,0 for control)}

\item{long.knot.nums}{number of knots for nonparametric curves (default 2 to 11)}
}
\value{
A list of analysis results

        1. monthly QOL results:

          - month: time variable

          - nCtr, nTrt: observed number of patients in control and treatment group, respectively

          - CtrQOL, TrtQOL: average monthly QOL in control and treatment group

          - Ctr95\%CIlow, Ctr95\%CIhi: lower and upper bounds of 95% CI for control group

          - Trt95\%CIlow, Trt95\%CIhi: lower and upper bounds of 95% CI for treatment group

          - TrtEffect: monthly treatment effect

          - pValueTrtEffect: p-value for monthly treatment effect

        2. The overall p value for testing the entire model

        3. knot.combinations: different combinations of knots for longitudinal and survival submodels with their corresponding AIC and BIC values

        4. maximum likelihood estimates (mle) results

        and

        Four figures saved in the current working directory

        1. estimated terminal trend in a png file called "Longitudinal_trajectories.png"

        2. the time-varying treatment effect in a png file called "Longitudinal_spline_treatment_effect.png"

        3. estimated survival function and in a png file "Survival_function.png"

        4. estimated cumulative hazard function in a png file "Cumulative_hazard.png"
}
\description{
Joint Model for Palliative care studies
}
\details{
A semiparametric joint model for terminal trend of quality of life (QOL) and survival

- For longitudinal QOL outcomes: semiparametric mixed effects submodel (only random intercept is used in this package)

\deqn{Y_i(t^*) = \beta_\mu(t^*) + A_i \beta_A(t^*) + X_i^T \psi_X + b_i + \epsilon_i(t^*)}

- For survival outcomes: Cox submodel
  \deqn{\lambda_i(t) = \lambda_0(t) exp(A_i\alpha_A + \tilde{X}_i^T \alpha_X)}
}
\examples{
data(dat_JMP)
results <- JMP(dat=dat_JMP,
covariate.fields = c("sex", "qol_0"),
qol.prefix="qol_", qol.time.prefix="time_",
id.field = "id", survival.time.field = "survival_time",
censoring.status.field = "death", treatment.status.field = "trt")
}
\references{
Li Z, Frost HR, Tosteson TD, et al. A Semiparametric Joint Model for Terminal Trend of Quality of Life and Survival in Palliative Care Research. Statistics in Medicine. 2017;36:4692–4704. https://doi.org/10.1002/sim.7445
}
\author{
Zhigang Li (zhigang.li @ ufl.edu)
}
